+++ CHANGELOG.md
@@ -37,6 +37,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Improvements
+
+* [#646](https://github.com/babylonlabs-io/finality-provider/pull/646) chore: adds more e2e for cosmos
+
 ## v2.0.0-rc.3
 
 ### Improvements